// src/app/core/models/disc-search-result.interface.ts
export interface DiscSearchResult {
    id: string;
    name: string;
    brand: string;
    category: string;
    speed: number;
    glide: number;
    turn: number;
    fade: number;
    stability: string;
    link: string;
    pic: string;
    name_slug: string;
    brand_slug: string;
    category_slug: string;
    stability_slug: string;
    color: string;
    background_color: string;
  }
  